fix(product): split oversized updater frontend files
Both files crossed the ~400-line frontend structure ceiling:
- use-updater.ts: extract the 11 action creators (check/download/retry/
  abort-first cancel/restart/prompt) into use-updater-actions.ts as
  useUpdaterActions(); public hook API unchanged
- UpdateUiPlaygroundControls.tsx: extract the phase-simulation mock
  builders and preview catalog into update-ui-playground-mocks.tsx

No behavior change; direct imports, no barrels.
