+++ apps/packages/product-client/src/hooks/access/tauri/use-updater.ts
@@ -1,28 +1,21 @@
-import { useCallback, useEffect, useMemo, useRef, useState } from "react";
+import { useEffect, useMemo, useRef, useState } from "react";
 import { useProductHost } from "@proliferate/product-client/host/ProductHostProvider";
 import { useUpdaterStore } from "#product/stores/updater/updater-store";
 import type { UpdaterErrorSource, UpdaterPhase } from "#product/stores/updater/updater-store";
 import { useProductTelemetry } from "#product/hooks/telemetry/facade/use-product-telemetry";
 import { useProductStorageContext } from "#product/hooks/persistence/facade/use-product-storage-context";
 import {
-  clearDevUpdaterMockDownload,
   DEV_UPDATER_MOCK_EVENT,
   isDevUpdaterMockSupported,
   readDevUpdaterMock,
   seedDevUpdaterMockFromEnv,
-  startDevUpdaterMockDownload,
-  updateDevUpdaterMock,
-  writeDevUpdaterMock,
   type DevUpdaterMockState,
 } from "./updater-dev-mock";
-import { abortOwnedDownload, runDownloadAndPrepareRestart } from "./updater-download";
 import {
   attachAutoCheckScheduler,
-  persistUpdaterMetadataSnapshot,
-  runUpdateCheck,
   type UpdaterSchedulerDeps,
 } from "./updater-check";
-import { readUpdaterFlags } from "./updater-flags";
+import { useUpdaterActions } from "./use-updater-actions";
 import { isDownloadStalled } from "#product/lib/domain/updates/download-stall";
 import { isOfficialHostedApiBaseUrl } from "#product/lib/infra/proliferate-api";
 
@@ -142,57 +135,19 @@ export function useUpdater() {
     };
   }, []);
 
-  const checkNow = useCallback(async () => {
-    if (devMock) {
-      const timestamp = new Date().toISOString();
-      const completedAt = Date.now();
-      updateDevUpdaterMock((current) =>
-        current
-          ? {
-              ...current,
-              lastCheckedAt: timestamp,
-              // Mirror the real flow: a manual check that finds no update raises
-              // the one-shot "up to date" signal.
-              manualCheckCompletedAt:
-                current.phase === "current" ? completedAt : current.manualCheckCompletedAt,
-            }
-          : current,
-      );
-      return;
-    }
-
-    if (!isPackaged || updater === null) {
-      return;
-    }
-    await runUpdateCheck(updater, depsRef.current, { userInitiated: true });
-  }, [devMock, isPackaged, updater]);
-
-  const clearManualCheckCompleted = useCallback(() => {
-    if (devMock) {
-      updateDevUpdaterMock((current) =>
-        current ? { ...current, manualCheckCompletedAt: null } : current,
-      );
-      return;
-    }
-    useUpdaterStore.getState().clearManualCheckCompleted();
-  }, [devMock]);
-
-  const downloadUpdate = useCallback(async () => {
-    if (devMock) {
-      startDevUpdaterMockDownload();
-      return;
-    }
-
-    if (!isPackaged || updater === null) {
-      return;
-    }
-    const owned = (await readUpdaterFlags(depsRef.current.storage))
-      .ownedUpdaterEnabled;
-    await runDownloadAndPrepareRestart(updater, depsRef.current, {
-      owned,
-      storage: depsRef.current.storage,
-    });
-  }, [devMock, isPackaged, updater]);
+  const {
+    checkNow,
+    clearManualCheckCompleted,
+    downloadUpdate,
+    retryDownload,
+    cancelUpdate,
+    skipVersion,
+    openRestartPrompt,
+    closeRestartPrompt,
+    scheduleRestartWhenIdle,
+    cancelRestartCountdown,
+    restartNow,
+  } = useUpdaterActions({ devMock, isPackaged, updater, depsRef, availableVersion });
 
   /**
    * Stall detection: while a download is in flight, poll the byte clock and
@@ -218,144 +173,6 @@ export function useUpdater() {
       window.clearInterval(interval);
     };
   }, [devMock, phase]);
-
-  const retryDownload = useCallback(async () => {
-    if (devMock) {
-      // "Retry now" is the stalled toast's commit button, so it has to do
-      // something under the mock too — otherwise the one recovery path out of
-      // the stall can't be exercised on the surface built to review it. Restart
-      // the forced download and re-arm the stall clock, mirroring the real
-      // store's `retryDownload`.
-      updateDevUpdaterMock((current) =>
-        current
-          ? {
-              ...current,
-              phase: "downloading",
-              downloadRetryCount: current.downloadRetryCount + 1,
-              lastProgressAt: Date.now(),
-              downloadStartedAt: Date.now(),
-            }
-          : current,
-      );
-      startDevUpdaterMockDownload();
-      return;
-    }
-
-    if (!isPackaged || updater === null) {
-      return;
-    }
-    // Abort-first: a retry must cancel any live (stalled) download and await the
-    // ack before starting a new one, so there is never more than one live
-    // download.
-    await abortOwnedDownload(updater);
-    const owned = (await readUpdaterFlags(depsRef.current.storage))
-      .ownedUpdaterEnabled;
-    useUpdaterStore.getState().retryDownload();
-    await runDownloadAndPrepareRestart(updater, depsRef.current, {
-      owned,
-      storage: depsRef.current.storage,
-    });
-  }, [devMock, isPackaged, updater]);
-
-  /** Abandon this update entirely; the pill and any toast go away. */
-  const cancelUpdate = useCallback(async () => {
-    if (devMock) {
-      writeDevUpdaterMock(null);
-      return;
-    }
-    // Abort-first: tear down any live owned download before resetting, so a
-    // reset can never strand a still-running transfer.
-    if (updater !== null) {
-      await abortOwnedDownload(updater);
-    }
-    useUpdaterStore.getState().reset();
-  }, [devMock, updater]);
-
-  const skipVersion = useCallback(() => {
-    const version = availableVersion;
-    if (devMock || version === null) {
-      return;
-    }
-    useUpdaterStore.getState().skipVersion(version);
-    // Persist so the skip survives relaunch (the store's skip list is seeded
-    // from this on the next boot).
-    void persistUpdaterMetadataSnapshot(depsRef.current.storage);
-  }, [availableVersion, devMock]);
-
-  const cancelRestartCountdown = useCallback(() => {
-    if (devMock) {
-      // "Not now" has to actually stop the countdown under the mock too, or the
-      // one cancellable state in the flow can't be exercised.
-      updateDevUpdaterMock((current) =>
-        current
-          ? { ...current, restartCountdownStartedAt: null, restartWhenIdle: false }
-          : current,
-      );
-      return;
-    }
-    useUpdaterStore.getState().cancelRestartCountdown();
-  }, [devMock]);
-
-  const openRestartPrompt = useCallback(() => {
-    if (devMock) {
-      updateDevUpdaterMock((current) =>
-        current ? { ...current, restartPromptOpen: true } : current,
-      );
-      return;
-    }
-    useUpdaterStore.getState().setRestartPromptOpen(true);
-  }, [devMock]);
-
-  const closeRestartPrompt = useCallback(() => {
-    if (devMock) {
-      updateDevUpdaterMock((current) =>
-        current ? { ...current, restartPromptOpen: false } : current,
-      );
-      return;
-    }
-    useUpdaterStore.getState().setRestartPromptOpen(false);
-  }, [devMock]);
-
-  const scheduleRestartWhenIdle = useCallback(() => {
-    if (devMock) {
-      updateDevUpdaterMock((current) =>
-        current ? { ...current, restartWhenIdle: true, restartPromptOpen: false } : current,
-      );
-      return;
-    }
-    const store = useUpdaterStore.getState();
-    store.setRestartWhenIdle(true);
-    store.setRestartPromptOpen(false);
-  }, [devMock]);
-
-  const restartNow = useCallback(async () => {
-    if (devMock) {
-      clearDevUpdaterMockDownload();
-      writeDevUpdaterMock(null);
-      return;
-    }
-
-    if (!isPackaged || updater === null) {
-      return;
-    }
-    useUpdaterStore.getState().setRestartPromptOpen(false);
-    // Owned path: `ready` means staged + verified, not yet installed. Install
-    // the staged bytes now (re-verified natively), then relaunch. The legacy
-    // path already installed during download, so it only relaunches.
-    const update = useUpdaterStore.getState()._update;
-    const owned = (await readUpdaterFlags(depsRef.current.storage))
-      .ownedUpdaterEnabled;
-    if (owned && update !== null && typeof updater.installStaged === "function") {
-      try {
-        await updater.installStaged(update);
-      } catch (error) {
-        const message = error instanceof Error ? error.message : String(error);
-        useUpdaterStore.getState().setError(message, "download");
-        return;
-      }
-    }
-    await updater.relaunch();
-  }, [devMock, isPackaged, updater]);
 
   useEffect(() => {
     if (!isPackaged || updater === null) {
